requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.3.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
httpx==0.25.2
selenium==4.15.0
webdriver-manager==4.0.1
python-multipart==0.0.6
Jinja2==3.1.2
